+++ springbootapi/springbootapi/src/main/java/com/max/back/neusoft/api/DrugServlet.java
@@ -3,6 +3,7 @@ package com.max.back.neusoft.api;
 import cn.hutool.core.bean.BeanUtil;
 import com.alibaba.fastjson.JSON;
 import com.alibaba.fastjson.serializer.SerializerFeature;
+import com.baomidou.mybatisplus.core.conditions.query.QueryWrapper;
 import com.baomidou.mybatisplus.core.conditions.update.UpdateWrapper;
 import com.max.back.common.http.ResponseResult;
 import com.max.back.neusoft.form.DeleteFrom;
@@ -10,9 +11,12 @@ import com.max.back.neusoft.form.DrugFindFrom;
 import com.max.back.neusoft.form.DrugModifyEnable;
 import com.max.back.neusoft.form.DrugSubmitFrom;
 import com.max.back.neusoft.pojo.Nondrug;
+import com.max.back.neusoft.pojo.Orderdrug;
 import com.max.back.neusoft.service.DrugspecificationService;
 import com.max.back.neusoft.service.NondrugService;
+import com.max.back.neusoft.service.OrderdrugService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Primary;
 import org.springframework.web.bind.annotation.*;
 import org.springframework.web.multipart.MultipartFile;
 
@@ -24,9 +28,8 @@ import java.io.ByteArrayOutputStream;
 import java.io.File;
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
-import java.util.List;
-import java.util.Map;
 import java.util.UUID;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 @CrossOrigin
 @RestController
@@ -40,6 +43,9 @@ public class DrugServlet {
     @Autowired
     private NondrugService nondrugService;
 
+    @Autowired
+    private OrderdrugService orderdrugService;
+
     //获取药品规格
     @GetMapping("/findAllSpecification")
     public String findAllSpecification() {
@@ -52,6 +58,11 @@ public class DrugServlet {
         return nondrugService.getAllDrugByRequire(drugFindFrom);
     }
 
+    /**
+     * 修改药品是否启用
+     * @param modifyEnable
+     * @return
+     */
     @PostMapping("/modifyEnable")
     public String modifyEnable(@RequestBody @Valid DrugModifyEnable modifyEnable) {
         UpdateWrapper<Nondrug> updateWrapper = new UpdateWrapper<>();
@@ -65,8 +76,25 @@ public class DrugServlet {
         }
     }
 
+    /**
+     * 删除药品
+     * @param params
+     * @return
+     */
     @PostMapping("/deleteByIds")
     public String deleteById(@RequestBody @Valid DeleteFrom params) {
+        //在订单表根据用户id查找是否存在相同用户
+        AtomicBoolean same = new AtomicBoolean(false);
+        params.getIds().stream().forEach(s -> {
+            QueryWrapper<Orderdrug> queryWrapper = new QueryWrapper<>();
+            queryWrapper.eq("o_drugId",s);
+            if(orderdrugService.getMap(queryWrapper)!=null){
+                same.set(true);
+            }
+        });
+        if(same.get()){
+            return JSON.toJSONString(ResponseResult.getErrorResult("C405"));
+        }
         boolean delete = nondrugService.removeByIds(params.getIds());
         if (delete) {
             return JSON.toJSONString(ResponseResult.getSuccessResult(null, "C200", null), SerializerFeature.DisableCircularReferenceDetect);
@@ -75,11 +103,21 @@ public class DrugServlet {
         }
     }
 
+    /**
+     * 上传药品封面
+     * @param file
+     */
     @PostMapping("/avatar")
     public void avatar(@RequestBody MultipartFile file) {
         multipartFile = file;
     }
 
+    /**
+     * 跟新药品信息
+     * @param drugSubmitFrom
+     * @return
+     * @throws IOException
+     */
     @PostMapping("/updateDrug")
     public String updateUser(@RequestBody DrugSubmitFrom drugSubmitFrom) throws IOException {
         if (multipartFile != null) {
@@ -109,6 +147,12 @@ public class DrugServlet {
         }
     }
 
+    /**
+     * 添加药品
+     * @param drugSubmitFrom
+     * @return
+     * @throws IOException
+     */
     @PostMapping("/addDrug")
     public String addDrug(@RequestBody @Valid DrugSubmitFrom drugSubmitFrom) throws IOException {
         if (multipartFile == null) {
@@ -135,6 +179,11 @@ public class DrugServlet {
         }
     }
 
+    /**
+     * 下载药品模板
+     * @param response
+     * @throws IOException
+     */
     @PostMapping("/dataTemplateDownload")
     public void userTemplateDownload(HttpServletResponse response) throws IOException {
         response.setContentType("application/vnd.openxmlformats-officedocument.spreadsheetml.sheet;charset=utf-8");
@@ -152,6 +201,12 @@ public class DrugServlet {
         outputStream.flush();
         outputStream.close();
     }
+
+    /**
+     * 下载药品excel全部数据
+     * @param response
+     * @throws IOException
+     */
     @PostMapping("/exportData")
     public void exportDataUser(HttpServletResponse response) throws IOException {
         response.setContentType("application/vnd.openxmlformats-officedocument.spreadsheetml.sheet;charset=utf-8");
@@ -169,12 +224,22 @@ public class DrugServlet {
         outputStream.flush();
         outputStream.close();
     }
+
+    /**
+     * 上传excel文件暴漏链接
+     * @param file
+     */
     @PostMapping("/uploadToServer")
     public void uploadToServer(@RequestBody MultipartFile file) {
         multipartFileExcel = file;
         fileName = UUID.randomUUID().toString();
     }
 
+    /**
+     * 对上传文件进行解析
+     * @return
+     * @throws IOException
+     */
     @GetMapping("/analyseFile")
     @ResponseBody
     public String analyseFile() throws IOException {
@@ -183,23 +248,39 @@ public class DrugServlet {
             File newFile = new File(path + "\\src\\main\\webapp\\static\\" + fileName + ".xlsx");
             multipartFileExcel.transferTo(newFile);
             return nondrugService.analyseFile(fileName);
-        }catch (NullPointerException e){
+        } catch (NullPointerException e) {
             return JSON.toJSONString(ResponseResult.getErrorResult("C502"));
         }
     }
+
+    /**
+     * 获取已逻辑删除药品信息
+     * @return
+     */
     @GetMapping("/getDeleteDrug")
-    public String getDeletes(){
-       return nondrugService.getDeletes();
+    public String getDeletes() {
+        return nondrugService.getDeletes();
     }
+
+    /**
+     * 永久删除
+     * @param id
+     * @return
+     */
     @PostMapping("/deletePermanently")
     public String deletePermanently(@RequestBody @NotBlank String id) {
         id = id.replace("=", "");
         String path = System.getProperty("user.dir");
-        File newFile = new File(path + "\\src\\main\\webapp\\Img\\nonDrug\\" + id+"\\"+id+".jpg");
+        File newFile = new File(path + "\\src\\main\\webapp\\Img\\nonDrug\\" + id + "\\" + id + ".jpg");
         newFile.delete();
         return nondrugService.deletePermanently(id);
     }
 
+    /**
+     * 恢复数据
+     * @param id
+     * @return
+     */
     @PostMapping("/dataRecovery")
     public String dataRecovery(@RequestBody @NotBlank String id) {
         id = id.replace("=", "");
